[mongoose] Use mongodb.UpdateQuery type
