+++ types/mongoose/tsconfig.json
@@ -18,6 +18,9 @@
     },
     "files": [
         "index.d.ts",
-        "mongoose-tests.ts"
+        "test/document.ts",
+        "test/model.ts",
+        "test/mongoose.ts",
+        "test/query.ts"
     ]
 }
